+++ oragono.yaml
@@ -341,8 +341,9 @@ debug:
 datastore:
     # path to the datastore
     path: ircd.db
-    # if the database schema requires an upgrade, `autoupgrade` (which defaults to true)
-    # will attempt to perform it automatically on startup. the database will be backed
+
+    # if the database schema requires an upgrade, `autoupgrade` will attempt to
+    # perform it automatically on startup. the database will be backed
     # up, and if the upgrade fails, the original database will be restored.
     autoupgrade: true
